+++ glade_forms/lib/src/widgets/widgets.dart
@@ -1,6 +1,6 @@
 export 'glade_form_builder.dart';
 export 'glade_form_consumer.dart';
 export 'glade_form_listener.dart';
-export 'glade_model_debug_info.dart';
-export 'glade_model_debug_info_modal.dart';
+export 'glade_form_debug_info.dart';
+export 'glade_form_debug_info_modal.dart';
 export 'glade_model_provider.dart';